refactor: move update helpers to upconv pkg
