--- telegram/internal/helpers/doc.go
@@ -1,2 +1,0 @@
-// Package helpers contains some common helping functions.
-package helpers
