--- telegram/internal/helpers/flood_wait.go
@@ -1,1 +1,0 @@
-package helpers
