+++ telegram/internal/upconv/doc.go
@@ -1,0 +1,2 @@
+// Package upconv contains updates conversion helpers.
+package upconv
